Create Jenkins installation
Using AWS CodePipeline, CodeBuild and CodeDeploy

Create a GitHub repository for your PHP application and push your code to it.
Set up an AWS CodePipeline that uses your GitHub repository as a source.
In the build stage of the pipeline, use AWS CodeBuild. Configure CodeBuild to use a build spec file
In the build spec file, specify commands to build your Docker image and push it to AWS Elastic Container Registry (ECR)
Add a deployment stage to your AWS CodePipeline that uses AWS CodeDeploy.
This stage will be triggered once the build stage completes successfully.
During this stage, AWS CodeDeploy will take the new Docker image, as specified in your AppSpec file, and run it on your EC2 instance.
You can use Jenkins as well if you are not familiar with Codepipeline.