Add files via upload
Task Details:

1. Create a PHP Web Application

Create a simple PHP application that interacts with a MySQL database.
Take any basic application from any github repository.
The application can be a basic form that takes user input and stores it in a MySQL database.
2. Dockerize the PHP Application

Create a Dockerfile for the application. This Dockerfile should use an official PHP Docker image and should copy your application files into the appropriate directory within the Docker image.
3. Set Up MySQL Database on AWS RDS

Create a MySQL database instance on AWS RDS. Connect your PHP application to this database and ensure it works as expected.
4. Manual Deployment to EC2

Create an AWS EC2 instance and install Docker on it.
Manually transfer your PHP application files and Dockerfile to the EC2 instance.
On the EC2 instance, build your Docker image using the docker build command and run it using the docker run command. Ensure the PHP application is able to connect and interact with the MySQL database hosted on AWS RDS.
5. Testing

Test your application by accessing the public IP of your EC2 instance. Verify that your application can read and write to the database.
6. Set Up CI\CD ( Use AWS Codepipeline or Jenkins):



Using AWS CodePipeline, CodeBuild and CodeDeploy

Create a GitHub repository for your PHP application and push your code to it.
Set up an AWS CodePipeline that uses your GitHub repository as a source.
In the build stage of the pipeline, use AWS CodeBuild. Configure CodeBuild to use a build spec file
In the build spec file, specify commands to build your Docker image and push it to AWS Elastic Container Registry (ECR)
Add a deployment stage to your AWS CodePipeline that uses AWS CodeDeploy.
This stage will be triggered once the build stage completes successfully.
During this stage, AWS CodeDeploy will take the new Docker image, as specified in your AppSpec file, and run it on your EC2 instance.
You can use Jenkins as well if you are not familiar with Codepipeline.